+++ photutils/background/interpolators.py
@@ -5,7 +5,6 @@ Define interpolator classes for Background2D.
 
 import numpy as np
 from astropy.units import Quantity
-from astropy.utils.decorators import deprecated_renamed_argument
 from scipy.ndimage import zoom
 
 from photutils.utils import ShepardIDWInterpolator
@@ -43,31 +42,23 @@ class BkgZoomInterpolator:
         input image. This is enabled by default, since higher order
         interpolation may produce values outside the given input range.
 
-    grid_mode : bool, optional
-        If `True` (default), the samples are considered as the centers
-        of regularly-spaced grid elements. If `False`, the samples
-        are treated as isolated points. For zooming 2D images,
-        this keyword should be set to `True`, which makes zoom's
-        behavior consistent with `scipy.ndimage.map_coordinates` and
-        `skimage.transform.resize`. The `False` option is provided only
-        for backwards-compatibility.
-
-        .. deprecated:: 2.0.0
-           When this keyword is removed, the behavior will be
-           ``grid_mode=True``.
+    Notes
+    -----
+    When resizing the mesh to the full image size, the samples are
+    considered as the centers of regularly-spaced grid elements (i.e.,
+    `~scipy.ndimage.zoom` ``grid_mode`` is True). This makes makes
+    zoom's behavior consistent with `scipy.ndimage.map_coordinates` and
+    `skimage.transform.resize`
     """
 
-    @deprecated_renamed_argument('grid_mode', None, '2.0.0')
-    def __init__(self, *, order=3, mode='reflect', cval=0.0, clip=True,
-                 grid_mode=True):
+    def __init__(self, *, order=3, mode='reflect', cval=0.0, clip=True):
         self.order = order
         self.mode = mode
         self.cval = cval
-        self.grid_mode = grid_mode
         self.clip = clip
 
     def __repr__(self):
-        params = ('order', 'mode', 'cval', 'clip', 'grid_mode')
+        params = ('order', 'mode', 'cval', 'clip')
         return make_repr(self, params)
 
     def __call__(self, data, **kwargs):
@@ -94,19 +85,13 @@ class BkgZoomInterpolator:
             return np.full(kwargs['shape'], np.min(data),
                            dtype=kwargs['dtype'])
 
-        if kwargs['edge_method'] == 'pad':
-            # The mesh is first resized to the larger padded-data size
-            # (i.e., zoom_factor should be an integer) and then cropped
-            # back to the final data size.
-            zoom_factor = kwargs['box_size']
-            result = zoom(data, zoom_factor, order=self.order, mode=self.mode,
-                          cval=self.cval, grid_mode=self.grid_mode)
-            result = result[0:kwargs['shape'][0], 0:kwargs['shape'][1]]
-        else:
-            # The mesh is resized directly to the final data size.
-            zoom_factor = np.array(kwargs['shape']) / data.shape
-            result = zoom(data, zoom_factor, order=self.order, mode=self.mode,
-                          cval=self.cval)
+        # The mesh is first resized to the larger padded-data size
+        # (i.e., zoom_factor should be an integer) and then cropped
+        # back to the final data size.
+        zoom_factor = kwargs['box_size']
+        result = zoom(data, zoom_factor, order=self.order, mode=self.mode,
+                      cval=self.cval, grid_mode=True)
+        result = result[0:kwargs['shape'][0], 0:kwargs['shape'][1]]
 
         if self.clip:
             minval = np.min(data)